requirements.txt
exa-py>=1.0.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.7.0
requests>=2.31.0
markdown>=3.4.0
pymupdf>=1.23.0
pyyaml>=6.0
